requirements.txt
flask
flask-wtf
flask-sqlalchemy
python-dotenv
argon2-cffi
cryptography